# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Audit Web IA API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/prospect": {
      "get": {
        "tags": [
          "prospect"
        ],
        "summary": "List Prospects",
        "operationId": "list_prospects_prospect_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProspectSheet"
                  },
                  "type": "array",
                  "title": "Response List Prospects Prospect Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      },
      "post": {
        "tags": [
          "prospect"
        ],
        "summary": "Start Prospect",
        "operationId": "start_prospect_prospect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProspectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProspectSheet"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/prospect/{prospect_id}": {
      "get": {
        "tags": [
          "prospect"
        ],
        "summary": "Get Prospect",
        "operationId": "get_prospect_prospect__prospect_id__get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "prospect_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prospect Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProspectSheet"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "prospect"
        ],
        "summary": "Delete Prospect",
        "operationId": "delete_prospect_prospect__prospect_id__delete",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "prospect_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prospect Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DetectedTech": {
        "properties": {
          "category": {
            "type": "string",
            "title": "Category"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "confidence": {
            "type": "string",
            "enum": [
              "high",
              "medium",
              "low"
            ],
            "title": "Confidence",
            "default": "medium"
          },
          "evidence": {
            "type": "string",
            "title": "Evidence",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "category",
          "name"
        ],
        "title": "DetectedTech",
        "description": "A single technology spotted on the prospect's site."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProspectCompanyIdentity": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "location": {
            "type": "string",
            "title": "Location",
            "default": ""
          },
          "sector": {
            "type": "string",
            "title": "Sector",
            "default": ""
          },
          "estimatedFoundedYear": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimatedfoundedyear"
          },
          "estimatedSize": {
            "type": "string",
            "title": "Estimatedsize",
            "default": ""
          },
          "socialProfiles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Socialprofiles"
          },
          "onlinePresenceNotes": {
            "type": "string",
            "title": "Onlinepresencenotes",
            "default": ""
          },
          "valueProposition": {
            "type": "string",
            "title": "Valueproposition",
            "default": ""
          }
        },
        "type": "object",
        "title": "ProspectCompanyIdentity",
        "description": "Who the company is — guessed from the site + light web search."
      },
      "ProspectPersona": {
        "properties": {
          "likelyContactRoles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Likelycontactroles"
          },
          "likelyPriorities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Likelypriorities"
          },
          "approachAngles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Approachangles"
          }
        },
        "type": "object",
        "title": "ProspectPersona",
        "description": "Likely decision-maker + tailored prospecting angles."
      },
      "ProspectRequest": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "ProspectRequest"
      },
      "ProspectSheet": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "done",
              "failed"
            ],
            "title": "Status",
            "default": "pending"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "identity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProspectCompanyIdentity"
              },
              {
                "type": "null"
              }
            ]
          },
          "stack": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProspectStackByCategory"
              },
              {
                "type": "null"
              }
            ]
          },
          "persona": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProspectPersona"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "url",
          "domain",
          "createdAt"
        ],
        "title": "ProspectSheet",
        "description": "A prospecting sheet job — request + status + result."
      },
      "ProspectStackByCategory": {
        "properties": {
          "cms": {
            "items": {
              "$ref": "#/components/schemas/DetectedTech"
            },
            "type": "array",
            "title": "Cms"
          },
          "analytics": {
            "items": {
              "$ref": "#/components/schemas/DetectedTech"
            },
            "type": "array",
            "title": "Analytics"
          },
          "advertising": {
            "items": {
              "$ref": "#/components/schemas/DetectedTech"
            },
            "type": "array",
            "title": "Advertising"
          },
          "chatCrm": {
            "items": {
              "$ref": "#/components/schemas/DetectedTech"
            },
            "type": "array",
            "title": "Chatcrm"
          },
          "hostingCdn": {
            "items": {
              "$ref": "#/components/schemas/DetectedTech"
            },
            "type": "array",
            "title": "Hostingcdn"
          },
          "other": {
            "items": {
              "$ref": "#/components/schemas/DetectedTech"
            },
            "type": "array",
            "title": "Other"
          }
        },
        "type": "object",
        "title": "ProspectStackByCategory",
        "description": "Detected tech stack, grouped by category."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}